Added build config, equip groups + names

+++ Insurgency/SFMLUIMenu.cpp
@@ -87,7 +87,7 @@ void SFMLUIMenu::moveCursorUp()
 }
 void SFMLUIMenu::moveCursorDown()
 {
-	if(m_selection ==  static_cast<int>(m_UIMenu->m_currentMenuList->m_options.size() - 1) || 
+	if(m_selection ==  static_cast<long int>(m_UIMenu->m_currentMenuList->m_options.size() - 1) || 
 			m_UIMenu->m_currentMenuList->m_options.size() == 1)
 		moveSelection(0);
 	else
